requirements.txt
SQLAlchemy==2.0.25
fastapi==0.109.0
pydantic==2.6.0
uvicorn==0.27.0
starlette==0.35.0
python-dotenv==1.0.1
aiosqlite==0.19.0
black==24.1.1
ruff==0.1.14
